+++ jx/recipes/dynamic.php
@@ -1,0 +1,59 @@
+<?php
+
+
+require_once(__DIR__ . "/../jxbase.php");
+require_once(__DIR__ . "/../findnode.php");
+
+
+/**
+ * Main dynamic JS -> XML recipe. Data drive conversions.
+ */
+
+class dynamic_js2xml extends J2X_Recipe {
+
+    public $recipe = null;
+
+    function __construct($recipe, $version='1.0', $encoding='UTF-8') {
+        parent::__construct(null, $version, $encoding);
+        $this->recipe = $recipe;
+    }
+
+    function topnode($node, $dom) {
+        if ($this->recipe === null) {
+            throw new NullData("no recipe provided");
+        }
+        foreach ($this->recipe as $instruction) {
+            list($datapath, $xpath, $extras) = $instruction;
+            // echo "datapath: $datapath xpath: $xpath extras: $extras\n";
+            $dnode = findnode($datapath, $node);
+            // var_dump($dnode);
+            if (is_scalar($dnode)) {
+                $xpathparts = pathparts($xpath);
+                $xpathtail = end($xpathparts);
+
+                if ($extras === "CDATA") {
+                    $dom->add_tree($xpath);
+                    $elt = $dom->xpathone($xpath);
+                    $cdata_node = new DOMCdataSection($dnode);
+                    $elt->appendChild($cdata_node);
+                }
+                elseif (startsWith($xpathtail, "@")) {
+                    $elt_xpathparts = array_slice($xpathparts, 0, count($xpathparts)-1);
+                    $elt_xpath = "/" . join('/', $elt_xpathparts);
+                    $attname = substr($xpathtail, 1, strlen($xpathtail)-1);
+                    $elt = $dom->add_tree($elt_xpath, null, array($attname => $dnode));
+                    // $dom->addAttrib($elt, $attname, $dnode);
+                }
+                else {
+                    $dom->add_tree($xpath, $dnode);
+                }
+            }
+            // echo "---\n";
+        }
+
+    }
+}
+
+// TODO: more richly handle [*] as part of specifications
+
+?>